pyproject.toml
[project]
name = "nlp-task"
version = "0.1.0"
description = "NLP task"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore<=3.1.1",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "boto3<=1.42.37",
    "celery>=5.6.2",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "google-generativeai>=0.8.6",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "socksio>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "celery-types>=0.24.0",
    "mypy>=1.19.1",
    "ruff>=0.14.14",
    "sqlalchemy[mypy]>=2.0.46",
]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
exclude = ["alembic"]
explicit_package_bases = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[[tool.mypy.overrides]]
module = "alembic.env"
ignore_errors = true

[tool.ruff]
line-length = 130
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi.params.Depends",
    "fastapi.Query",
    "fastapi.Path",
    "fastapi.Body",
    "fastapi.Security",
    "fastapi.Header",
    "fastapi.Cookie",
    "fastapi.Form",
    "fastapi.File",
]
